wip: make sound
