+++ package.json
@@ -12,6 +12,7 @@
     "next": "10.2.3",
     "react": "17.0.2",
     "react-dom": "17.0.2",
-    "react-p5-wrapper": "^2.4.1"
+    "react-p5-wrapper": "^2.4.1",
+    "tone": "^14.7.77"
   }
 }
